third_party/blink/web_tests/platform/generic/fast/css/content-language-dynamically-changed-expected.txt

Test for bug 76701: map http-equiv content-language to -webkit-locale. This particular test tests that dynamic changes to the meta element have no effect. This expectation may change, see bug. The HTML 5 spec decrees that the pragma-set default language be set only when the meta element is inserted into the document. In Firefox and IE changing the element dynamically doesn't seem to affect the page, but in Firefox the meta content-language displayed in "Page Information" is affected.

PASS languageOfNode('x') is "\"zh\""
PASS languageOfNode('y') is "\"ar\""
FAIL languageOfNode('x') should be "zh". Was "ja".
PASS languageOfNode('y') is "\"ar\""
FAIL languageOfNode('x') should be "zh". Was "ja".
PASS languageOfNode('y') is "\"ar\""
FAIL languageOfNode('x') should be "zh". Was "ja".
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
